Refactor JobCompletionDetectAlgorithm as generic interface amd rename
