+++ shardingsphere-test/shardingsphere-pipeline-test/src/test/resources/META-INF/services/org.apache.shardingsphere.data.pipeline.spi.detect.JobCompletionDetectAlgorithm
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.data.pipeline.core.fixture.FixtureRuleAlteredJobCompletionDetectAlgorithm
+org.apache.shardingsphere.data.pipeline.core.fixture.FixtureJobCompletionDetectAlgorithm
